Update to allow this to work in OSX with Safari.
Without an html extension, Safari will not open the resulting file.

Also, in testing if the browser exists, it doesn't allow the addition of
arguments on the call.  Thus, there is also a change to the way
existance of executable is checked.  Splits the browser string and
checks the first item.