Calling the view program was not properly redirected.
There was some missing redirection code.  Here is how the current commit
redirects output.

When searching for an app to run all output is redirected to /dev/null
as the ruby "system" command return true/false based on the return code
fo the application.  Since we are looking at the 'which' command it will
return 0 (true) if it finds the app and 1 (false) if it doesn't.

When executing the viewer, stderr is redirected to stdout and stdout is
redirected to /dev/null.  Note that stderr is _not_ redirected to
/dev/null because of order of redirection.  However, this has the nice
consequence of printing an error in the VIM window if the preview
application cannot be called or has any err output.