Another fix for programs with arguments.
exec wants either multiple arguments, i.e., exec "prog", "arg1", "arg2".
OR it wants a single string "prog arg1 arg2 ...".  On OSX and when you
want a program to use arguments when loading, you need to consider that
the value of `app` is not just a single word (representing an
application).  In my case, `app="open -g -a Safari" because of OSX and
because I want it to load in the background (-g).